fix jacobian concatenate bug with functions R -> R^n for n > 1, add tests and move tests to their own file
